+++ test/runtime-puppeteer/samples/transition-css-out-in/main.svelte
@@ -1,0 +1,20 @@
+<script>
+	export let visible;
+
+	function foo() {
+		return {
+			duration: 10,
+			css: t => {
+				return `opacity: ${t}`;
+			}
+		};
+	}
+</script>
+
+{#if visible}
+	<div transition:foo></div>
+{/if}
+
+{#if !visible}
+	<div transition:foo></div>
+{/if}
