--- test/runtime/samples/style_manager-cleanup/main.svelte
@@ -1,14 +1,0 @@
-<script>
-    export let visible = false;
-    
-    function foo() {
-        return {
-            duration: 100,
-            css: () => ''
-        };
-    }
-    </script>
-    
-    {#if visible}
-        <div transition:foo></div>
-    {/if}
